created a login form component and separeted the logic from the page.tsx
